# backend/app/admin/service/cmd/server/assets/openapi.yaml
...
openapi: 3.0.3
info:
    title: Kratos-CMS 后台管理API
    description: Kratos-CMS 后台管理API
    contact:
        name: tx7do
        url: https://github.com/tx7do/kratos-cms
        email: yanglinbo@gmail.com
    license:
        name: MIT License
        url: https://github.com/tx7do/kratos-cms/blob/master/LICENSE
    version: "1.0"
paths:
...
    /admin/v1/tags:
        get:
            tags:
                - TagService
            description: 获取标签列表
            operationId: TagService_ListTag
            parameters:
                - name: page
                  in: query
                  description: 当前页码
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  description: 每页的行数
                  schema:
                    type: integer
                    format: int32
                - name: query
                  in: query
                  description: 与过滤参数
                  schema:
                    type: string
                - name: or
                  in: query
                  description: 或过滤参数
                  schema:
                    type: string
                - name: orderBy
                  in: query
                  description: 排序条件
                  schema:
                    type: array
                    items:
                        type: string
                - name: nopaging
                  in: query
                  description: 是否不分页
                  schema:
                    type: boolean
                - name: field_mask
                  in: query
                  description: 字段掩码
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListTagResponse'
        post:
            tags:
                - TagService
            description: 创建标签
            operationId: TagService_CreateTag
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateTagRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Tag'
    /admin/v1/tags/{id}:
        get:
            tags:
                - TagService
            description: 获取标签数据
            operationId: TagService_GetTag
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Tag'
        put:
            tags:
                - TagService
            description: 更新标签
            operationId: TagService_UpdateTag
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
                - name: operatorId
                  in: query
                  schema:
                    type: integer
                    format: uint32
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Tag'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Tag'
        delete:
            tags:
                - TagService
            description: 删除标签
            operationId: TagService_DeleteTag
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
                - name: operatorId
                  in: query
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content: {}
...
components:
    schemas:
        Any:
            type: object
            properties:
                value:
                    $ref: '#/components/schemas/GoogleProtobufAny'
                yaml:
                    type: string
...
        CreateTagRequest:
            type: object
            properties:
                tag:
                    $ref: '#/components/schemas/Tag'
                operatorId:
                    type: integer
                    format: uint32
            description: 请求 - 创建标签
...
        ExternalDocs:
            type: object
            properties:
                description:
                    type: string
                url:
                    type: string
                specificationExtension:
                    type: array
                    items:
                        $ref: '#/components/schemas/NamedAny'
            description: Allows referencing an external resource for extended documentation.
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
...
        ListTagResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/Tag'
                total:
                    type: integer
                    format: int32
            description: 回应 - 标签列表
...
        NamedAny:
            type: object
            properties:
                name:
                    type: string
                    description: Map key
                value:
                    allOf:
                        - $ref: '#/components/schemas/Any'
                    description: Mapped value
            description: Automatically-generated message used to represent maps of Any as ordered (name,value) pairs.
...
        Tag:
            type: object
            properties:
                name:
                    type: string
                description:
                    type: string
                externalDocs:
                    $ref: '#/components/schemas/ExternalDocs'
                specificationExtension:
                    type: array
                    items:
                        $ref: '#/components/schemas/NamedAny'
            description: Adds metadata to a single tag that is used by the Operation Object. It is not mandatory to have a Tag Object per tag defined in the Operation Object instances.
...
    securitySchemes:
        OAuth2PasswordBearer:
            type: oauth2
            flows:
                password:
                    tokenUrl: /admin/v1/login
                    refreshUrl: /admin/v1/refresh_token
                    scopes: {}
security:
    - OAuth2PasswordBearer: []
...